layout/tests: Extract fullscreen tests to submodule
